ML project regarding to Sign Language Recognition
Sign language has a long history in Western society as a visual language or technique of
communication in the olden days. Traditional gestures, hand signs, and figure spelling, as well
as the use of hand position to represent letters of the alphabet, make it a sign language. The
machine learning-based sign language recognition system is designed to communicate with
people with disabilities where it translates captured characters or gestures into text or audio so
that the system can easily read and know what the person is going to convey, whether or not
the user has the knowledge of sign languages.
The advancement in technology is to help individuals who are disabled. Sign language plays a
major role in communicating the needs of the user's information. The website captures the
gestures and displays the outcome through audio notification. Machine learning and other
technologies are used for the development of the project. The main focus of this work is to
adopt the hand gestures where the sensor is used to detect in a website so that it can identify
the needs of disabled people. This project aims to motivate the lives of the world by giving the
path with human interactions between websites where it provides a quick response to the users.